Move navbar to routes folder
