+++ client/src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css'
 import 'bootstrap/dist/css/bootstrap.min.css';
-import MyNavBar from './components/Navbar'
+import MyNavBar from './routes/Navbar'
 import ListStudents from './components/ListStudents'
